refactor: template 파일 -> 연관성 있는 view로 나누기
